refactor: 도메인 모델을 도메인별 하위 패키지로 분리
- model/ 하위를 auth, video, tripplan, place, common 패키지로 그룹화
- video·tripplan에서 공유하는 place 모델은 독립 패키지로 분리
- 범용 페이지네이션 타입 CursorPage는 common에 배치
- domain README의 디렉토리 구조·모델 작성 규칙 갱신
