+++ data/src/commonMain/kotlin/com/linkit/company/data/mapper/AuthMapper.kt
@@ -1,7 +1,7 @@
 package com.linkit.company.data.mapper
 
 import com.linkit.company.data.dto.auth.AuthResponse
-import com.linkit.company.domain.model.Auth
+import com.linkit.company.domain.model.auth.Auth
 
 internal fun AuthResponse.toDomain(): Auth {
     return Auth(
