--- domain/src/commonMain/kotlin/com/linkit/company/domain/model/.gitkeep
@@ -1,1 +1,0 @@
-# Domain models will be placed here
