+++ domain/src/commonMain/kotlin/com/linkit/company/domain/model/auth/Auth.kt
@@ -1,4 +1,4 @@
-package com.linkit.company.domain.model
+package com.linkit.company.domain.model.auth
 
 data class Auth(
     val memberId: String,
