+++ domain/src/commonMain/kotlin/com/linkit/company/domain/model/place/Place.kt
@@ -1,4 +1,4 @@
-package com.linkit.company.domain.model
+package com.linkit.company.domain.model.place
 
 data class Place(
     val id: String,
