+++ domain/src/commonMain/kotlin/com/linkit/company/domain/model/place/PlaceStatus.kt
@@ -1,4 +1,4 @@
-package com.linkit.company.domain.model
+package com.linkit.company.domain.model.place
 
 /** 일정 아이템의 Google Places 정보 연결 상태 */
 enum class PlaceStatus {
