+++ domain/src/commonMain/kotlin/com/linkit/company/domain/model/tripplan/TripPlanItemOrder.kt
@@ -1,4 +1,4 @@
-package com.linkit.company.domain.model
+package com.linkit.company.domain.model.tripplan
 
 /** 여행 계획 수정 시 아이템의 일차(day)·순서(itemOrder) 재배치 명령 */
 data class TripPlanItemOrder(
