+++ domain/src/commonMain/kotlin/com/linkit/company/domain/model/video/RecentVideo.kt
@@ -1,4 +1,4 @@
-package com.linkit.company.domain.model
+package com.linkit.company.domain.model.video
 
 /** 채널의 최신 영상 요약 */
 data class RecentVideo(
